backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yfinance>=1.2.0
scipy==1.14.1
numpy==2.1.1
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
